requirements.txt
pytest==7.1.1
selenium==4.1.3
webdriver-manager==3.5.4
allure-pytest=2.9.45